fix: [SCRUM-62] Using memo to prevent re-rendering when props haven't changed and Extract formatTime function as a utils to reuse
